update schema
changes:
- Format for links changed from "[location](#Location)" to {@link Location location}.
Pyright don't support the new format,
but the old one also didn't work as expected because it opened a blank browser window.

- Introduction of a new property "deprecated".
I didn't implement the display for "deprecated", because we already display text that a property is being deprecated.
Just search for the word "@deprecated" in lsp_types.py and you will see that we already display the deprecated text.
If I were to implement it, which I did, but reverted, I saw duplicated text.

- they reordered some types. don't know why.

- some types changed, "ConfigurationParams"